Move bottomsheet menu to gamemenu.dart
